+++ src/classes/Approvals.ts
@@ -1,23 +1,39 @@
 import OneBlinkAPI from '../lib/one-blink-api'
-import { SubmissionTypes, ApprovalTypes } from '@oneblink/types'
+import { SubmissionTypes, ApprovalTypes, FormTypes } from '@oneblink/types'
 import { Tenant, ConstructorOptions } from '../lib/types'
 
+export type FormSubmissionApprovalHistoryRecord = {
+  formSubmissionMeta: SubmissionTypes.FormSubmissionMeta
+  formApprovalFlowInstance: ApprovalTypes.FormApprovalFlowInstance
+  formSubmissionApprovals: ApprovalTypes.FormSubmissionApproval[]
+}
+
 export type FormSubmissionsAdministrationApprovalsResponse = {
-  approvals: Array<{
-    formSubmissionMeta: SubmissionTypes.FormSubmissionMeta
-    formApprovalFlowInstance: ApprovalTypes.FormApprovalFlowInstance
-    formSubmissionApprovals: ApprovalTypes.FormSubmissionApproval[]
-    history: Array<{
-      formSubmissionMeta: SubmissionTypes.FormSubmissionMeta
-      formApprovalFlowInstance: ApprovalTypes.FormApprovalFlowInstance
-      formSubmissionApprovals: ApprovalTypes.FormSubmissionApproval[]
-    }>
-  }>
+  approvals: Array<
+    FormSubmissionApprovalHistoryRecord & {
+      history: FormSubmissionApprovalHistoryRecord[]
+    }
+  >
   meta: {
     limit?: number
     offset?: number
     nextOffset?: number
   }
+}
+
+export type FormSubmissionApprovalResponse = {
+  formSubmissionMeta: SubmissionTypes.FormSubmissionMeta
+  formSubmissionApproval: ApprovalTypes.FormSubmissionApproval
+  formApprovalFlowInstance: ApprovalTypes.FormApprovalFlowInstance
+  form: FormTypes.Form
+  history: FormSubmissionApprovalHistoryRecord[]
+}
+
+export type FormApprovalFlowInstanceResponse = {
+  formSubmissionMeta: SubmissionTypes.FormSubmissionMeta
+  formApprovalFlowInstance: ApprovalTypes.FormApprovalFlowInstance
+  form: FormTypes.Form
+  formSubmissionApprovals: ApprovalTypes.FormSubmissionApproval[]
 }
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
@@ -28,7 +44,7 @@ export default (tenant: Tenant) =>
       super(options.accessKey, options.secretKey, tenant)
     }
 
-    async getFormSubmissionAdministrationApprovals({
+    async searchFormSubmissionAdministrationApprovals({
       formsAppId,
       limit,
       offset,
@@ -58,5 +74,23 @@ export default (tenant: Tenant) =>
         offset,
         ...rest,
       })
+    }
+
+    async getFormSubmissionApproval(
+      id: string,
+    ): Promise<FormSubmissionApprovalResponse> {
+      if (!id) {
+        throw new Error('"id" must be a string and is required')
+      }
+      return await super.getRequest(`/form-submission-approvals/${id}`)
+    }
+
+    async getFormApprovalFlowInstance(
+      id: number,
+    ): Promise<FormApprovalFlowInstanceResponse> {
+      if (typeof id !== 'number') {
+        throw new Error('"id" must be a number and is required')
+      }
+      return await super.getRequest(`/form-approval-flow-instances/${id}`)
     }
   }