+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [`getFormSubmissionApproval()`](./docs/approvals.md#getformsubmissionapproval) function
+- [`getFormApprovalFlowInstance()`](./docs/approvals.md#getformapprovalflowinstance) function
+
+### Renamed
+
+- [`searchFormSubmissionAdministrationApprovals()`](./docs/approvals.md#searchformsubmissionadministrationapprovals) function
+
 ## [0.11.2] - 2021-09-09
 
 ### Fixed